requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
python-dotenv
openpyxl
pydantic[email]
passlib[bcrypt]
python-jose
